feat: auth
