+++ .env.example
@@ -25,6 +25,6 @@ CLICKHOUSE_PASSWORD=password
 CLICKHOUSE_DB=default
 
 #* Authentication
-ADMIN_EMAIL=test@mail.com
+ADMIN_EMAIL=admin@example.com
 ADMIN_PASSWORD=secret
 JWT_SECRET=secret
