+++ apps/app/src/app/api/jobs/stats/schemas.ts
@@ -1,5 +1,5 @@
 import z from "zod";
-import { registerApiRoute } from "~/lib/utils/server";
+import { registerApiRoute } from "~/lib/utils/client";
 
 export const getJobsStatsInput = z.object({
   days: z.number().min(1).max(30),
