+++ apps/app/src/app/api/queues/stats/schemas.ts
@@ -1,5 +1,5 @@
 import z from "zod";
-import { registerApiRoute } from "~/lib/utils/server";
+import { registerApiRoute } from "~/lib/utils/client";
 
 export const getQueuesStatsOutput = z.object({
   total: z.number(),
