+++ apps/app/src/app/api/queues/table/schemas.ts
@@ -1,5 +1,5 @@
 import z from "zod";
-import { registerApiRoute } from "~/lib/utils/server";
+import { registerApiRoute } from "~/lib/utils/client";
 
 export const getQueuesTableInput = z.object({
   cursor: z.string().nullish(),
