+++ apps/app/src/app/layout.tsx
@@ -1,6 +1,7 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
+import { headers } from "next/headers";
 import { AuthGuard } from "~/components/auth-guard";
 import { Toaster } from "~/components/ui/sonner";
 import { Providers } from "./providers";
@@ -20,18 +21,20 @@ export const metadata: Metadata = {
   description: "A board to monitor your bullmq jobs",
 };
 
-export default function RootLayout({
+export default async function RootLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
+  const pathname = (await headers()).get("x-pathname") as string;
+
   return (
     <html lang="en">
       <body
         className={`${geistSans.variable} ${geistMono.variable} antialiased`}
       >
         <Providers>
-          <AuthGuard>{children}</AuthGuard>
+          <AuthGuard pathname={pathname}>{children}</AuthGuard>
         </Providers>
         <Toaster />
       </body>
