+++ apps/app/src/app/login/page.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-import { useRouter } from "next/navigation";
 import { useId, useState } from "react";
 import { Button } from "~/components/ui/button";
 import { Card } from "~/components/ui/card";
@@ -11,7 +10,6 @@ export default function LoginPage() {
   const [password, setPassword] = useState("");
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState("");
-  const router = useRouter();
 
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
@@ -30,8 +28,7 @@ export default function LoginPage() {
       const data = await response.json();
 
       if (data.success) {
-        router.push("/");
-        router.refresh();
+        window.location.href = "/";
       } else {
         setError(data.error || "Login failed");
       }
