+++ apps/app/src/app/queues/_components/queues-table.tsx
@@ -15,7 +15,7 @@ import {
   TableHeader,
   TableRow,
 } from "~/components/ui/table";
-import { apiFetch, cn } from "~/lib/utils/server";
+import { apiFetch, cn } from "~/lib/utils/client";
 import { QueueActions } from "./queue-actions";
 import { type TimePeriod, TimePeriodSelector } from "./time-period-selector";
