+++ apps/app/src/app/runs/_components/runs-filters.tsx
@@ -14,7 +14,7 @@ import {
   PopoverContent,
   PopoverTrigger,
 } from "~/components/ui/popover";
-import { apiFetch } from "~/lib/utils/server";
+import { apiFetch } from "~/lib/utils/client";
 import type { TRunFilters } from "./types";
 
 export function RunsFilters({
