+++ apps/app/src/components/ui/card.tsx
@@ -1,6 +1,6 @@
 import type * as React from "react";
 
-import { cn } from "~/lib/utils/server";
+import { cn } from "~/lib/utils/client";
 
 function Card({ className, ...props }: React.ComponentProps<"div">) {
   return (
