+++ apps/app/src/components/ui/combobox.tsx
@@ -16,7 +16,7 @@ import {
   PopoverContent,
   PopoverTrigger,
 } from "~/components/ui/popover";
-import { cn } from "~/lib/utils/server";
+import { cn } from "~/lib/utils/client";
 import { Loader } from "./loader";
 
 export type ComboboxOption = {
