+++ apps/app/src/components/ui/input.tsx
@@ -1,6 +1,6 @@
 import type * as React from "react";
 
-import { cn } from "~/lib/utils/server";
+import { cn } from "~/lib/utils/client";
 
 function Input({ className, type, ...props }: React.ComponentProps<"input">) {
   return (
