+++ apps/app/src/components/ui/popover.tsx
@@ -3,7 +3,7 @@
 import * as PopoverPrimitive from "@radix-ui/react-popover";
 import type * as React from "react";
 
-import { cn } from "~/lib/utils/server";
+import { cn } from "~/lib/utils/client";
 
 function Popover({
   ...props
