+++ apps/app/src/components/ui/select.tsx
@@ -4,7 +4,7 @@ import * as SelectPrimitive from "@radix-ui/react-select";
 import { CheckIcon, ChevronDownIcon, ChevronUpIcon } from "lucide-react";
 import type * as React from "react";
 
-import { cn } from "~/lib/utils/server";
+import { cn } from "~/lib/utils/client";
 
 function Select({
   ...props
