+++ apps/app/src/lib/auth/client.ts
@@ -1,0 +1,1 @@
+export const COOKIE_NAME = "bbb-token";
