+++ apps/app/src/lib/auth/server.ts
@@ -2,9 +2,9 @@ import { jwtVerify, SignJWT } from "jose";
 import type { ResponseCookies } from "next/dist/compiled/@edge-runtime/cookies";
 import { cookies } from "next/headers";
 import { env } from "../env";
+import { COOKIE_NAME } from "./client";
 
 const JWT_SECRET = new TextEncoder().encode(env.JWT_SECRET);
-const COOKIE_NAME = "auth-token";
 
 export interface User {
   email: string;
@@ -61,12 +61,8 @@ export async function verifyToken(token: string): Promise<User | null> {
 export async function getTokenFromRequest(): Promise<string | null> {
   const cookieStore = await cookies();
   // Try Authorization header first
-  const authHeader = cookieStore.get("authorization");
-  if (authHeader?.value?.startsWith("Bearer ")) {
-    return authHeader.value.slice(7);
-  }
-
-  return null;
+  const authHeader = cookieStore.get(COOKIE_NAME);
+  return authHeader?.value ?? null;
 }
 
 /**
