+++ apps/app/src/middleware.ts
@@ -1,0 +1,20 @@
+import { type NextRequest, NextResponse } from "next/server"
+
+export async function middleware(request: NextRequest) {
+  const response = NextResponse.next()
+  // Inject the current pathname in the headers
+  response.headers.set("x-pathname", request.nextUrl.pathname)
+
+  return response
+}
+
+export const config = {
+  /*
+   * Match all request paths except for the ones starting with:
+   * - api (API routes)
+   * - _next/static (static files)
+   * - _next/image (image optimization files)
+   * - favicon.ico (favicon file)
+   */
+  matcher: ["/((?!api|_next/static|_next/image|favicon.ico|public).*)"],
+}
